# swagger.yaml
openapi: 3.0.0
info:
  title: Ferramentas WEB API
  description: API REST para gerenciamento de produtos da loja de ferramentas Ferramentas WEB. Desenvolvido para a disciplina de Desenvolvimento Full Stack Básico.
  version: 1.0.0

servers:
  - url: /
    description: Servidor Atual

paths:
  /api/produtos:
    get:
      summary: Lista todos os produtos cadastrados
      description: Retorna uma lista com todos os produtos presentes no banco de dados SQLite.
      responses:
        '200':
          description: Lista de produtos retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

    post:
      summary: Cadastra um novo produto
      description: Recebe dados JSON e insere um novo produto no banco de dados.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - category
              properties:
                name:
                  type: string
                  example: Martelo de Garra
                price:
                  type: number
                  format: float
                  example: 35.90
                category:
                  type: string
                  example: Ferramentas Manuais
                description:
                  type: string
                  example: Martelo de ferro com cabo de madeira
      responses:
        '201':
          description: Produto cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Dados inválidos fornecidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /api/produtos/{id}:
    get:
      summary: Busca um produto pelo ID
      description: Retorna os detalhes de um único produto específico com base no ID fornecido na URL.
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do produto
          schema:
            type: integer
      responses:
        '200':
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

    put:
      summary: Atualiza um produto existente
      description: Modifica os dados de um produto já cadastrado baseado no ID. Todos os campos obrigatórios devem ser enviados.
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do produto a ser atualizado
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - price
                - category
              properties:
                name:
                  type: string
                  example: Martelo de Garra Cabo Longo
                price:
                  type: number
                  format: float
                  example: 45.00
                category:
                  type: string
                  example: Ferramentas Manuais
                description:
                  type: string
                  example: Martelo cabo longo de fibra de vidro
      responses:
        '200':
          description: Produto atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        '400':
          description: Requisição inválida
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

    delete:
      summary: Remove um produto pelo ID
      description: Exclui permanentemente um produto do banco de dados com base no ID.
      parameters:
        - name: id
          in: path
          required: true
          description: ID único do produto a ser deletado
          schema:
            type: integer
      responses:
        '200':
          description: Produto removido com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  mensagem:
                    type: string
                    example: Produto removido com sucesso.
                  id:
                    type: integer
                    example: 1
        '404':
          description: Produto não encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /api/categorias:
    get:
      summary: Lista todas as categorias cadastradas
      description: Retorna uma lista com todas as categorias presentes na tabela do banco de dados SQLite.
      responses:
        '200':
          description: Lista de categorias retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Categoria'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

  /api/produtos/busca:
    get:
      summary: Pesquisa produtos pelo nome
      description: Retorna produtos cujo nome corresponda parcialmente ao termo de busca fornecido (utilizando LIKE no SQL).
      parameters:
        - name: nome
          in: query
          required: true
          description: Termo ou parte do nome para buscar
          schema:
            type: string
      responses:
        '200':
          description: Resultados da busca
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'
        '400':
          description: Parâmetro 'nome' ausente na busca
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Erro'

components:
  schemas:
    Produto:
      type: object
      properties:
        id:
          type: integer
          description: ID do produto gerado pelo banco
          example: 1
        name:
          type: string
          description: Nome do produto/ferramenta
          example: Martelo de Garra
        price:
          type: number
          format: float
          description: Preço unitário do produto
          example: 35.90
        category:
          type: string
          description: Categoria à qual pertence o produto
          example: Ferramentas Manuais
        description:
          type: string
          description: Descrição curta do produto
          example: Martelo de ferro com cabo de madeira
        created_at:
          type: string
          description: Data e hora do cadastro do produto
          example: "2026-07-04 12:00:00"

    Categoria:
      type: object
      properties:
        id:
          type: integer
          description: ID da categoria gerado pelo banco
          example: 1
        name:
          type: string
          description: Nome da categoria
          example: Ferramentas Manuais

    Erro:
      type: object
      properties:
        erro:
          type: string
          description: Mensagem descritiva do erro
          example: Produto não encontrado.